Refactors vertical orbit implementation
Consolidates vertical orbit functionality into a single class.

This change removes the `VerticalLyapunovOrbit` class and introduces a `VerticalOrbit` class, simplifying the code and improving maintainability. The new `VerticalOrbit` class encompasses the behavior of the previous class, streamlining the implementation of vertical orbit calculations.

Refactors vertical orbit implementation

Consolidates vertical orbit functionality into a single class.

This change removes the `VerticalLyapunovOrbit` class and
introduces a `VerticalOrbit` class, simplifying the code and
improving maintainability. The new `VerticalOrbit` class
encompasses the behavior of the previous class, streamlining
the implementation of vertical orbit calculations.
