+++ src/hiten/system/libration/base.py
@@ -480,9 +480,10 @@ class LibrationPoint(ABC):
         """
 
         # Lazy imports to avoid circular dependencies and reduce import time.
-        from hiten.system.orbits.halo import HaloOrbit
-        from hiten.system.orbits.lyapunov import LyapunovOrbit, VerticalLyapunovOrbit
         from hiten.system.orbits.base import GenericOrbit, PeriodicOrbit
+        from hiten.system.orbits.halo import HaloOrbit
+        from hiten.system.orbits.lyapunov import LyapunovOrbit
+        from hiten.system.orbits.vertical import VerticalOrbit
 
         # Direct class provided
         if isinstance(family, type) and issubclass(family, PeriodicOrbit):
@@ -497,8 +498,8 @@ class LibrationPoint(ABC):
         mapping: dict[str, type[PeriodicOrbit]] = {
             "halo": HaloOrbit,
             "lyapunov": LyapunovOrbit,
-            "vertical_lyapunov": VerticalLyapunovOrbit,
-            "vertical": VerticalLyapunovOrbit,
+            "vertical_lyapunov": VerticalOrbit,
+            "vertical": VerticalOrbit,
             "generic": GenericOrbit,
         }
