+++ src/hiten/system/orbits/__init__.py
@@ -7,14 +7,10 @@ Usage example::
     from hiten.system.orbits import HaloOrbit, LyapunovOrbit
 """
 
-from .base import (
-    _CorrectionConfig,
-    PeriodicOrbit,
-    GenericOrbit,
-    S,
-)
+from .base import GenericOrbit, PeriodicOrbit, S, _CorrectionConfig
 from .halo import HaloOrbit
-from .lyapunov import LyapunovOrbit, VerticalLyapunovOrbit
+from .lyapunov import LyapunovOrbit
+from .vertical import VerticalOrbit
 
 __all__ = [
     "_CorrectionConfig",
@@ -22,6 +18,6 @@ __all__ = [
     "GenericOrbit",
     "HaloOrbit",
     "LyapunovOrbit",
-    "VerticalLyapunovOrbit",
+    "VerticalOrbit",
     "S",
 ]
